requirements.txt
gym==0.19.0
matplotlib==3.5.0
networkx==2.6.3
opencv_python==4.5.4.60
Pillow==8.4.0
torch_geometric==2.0.2
wandb==0.12.7
torch~=1.10.1
torch_scatter==2.0.9
torch_sparse==0.6.12
numpy==1.21.4
stable-baselines3==1.3.0
jupyter==1.0.0
algorithmx==2.0.3


dill~=0.3.4
opencv-python~=4.5.4.60